Provide explicit up to date python submodule structure
This rids of us the warning:

Couldn't find the symbol `PyInit_autopy` in the native library. Python will fail to import this module. If you're using pyo3, check that `#[pymodule]` uses `autopy` as module name

and eventual follow-up errors in importing exactly as warned and
provides up to date explicit python submodule structure for the
current PyO3 version.